+++ test/e2e/configurer/factory.go
@@ -31,6 +31,8 @@ type Configurer interface {
 	RunCosmosRelayerIBC() error
 
 	RunIBCTransferChannel() error
+	// OpenZoneConciergeChannel configures IBC channel using zoneconcierge port
+	OpenZoneConciergeChannel(chainA, chainB *chain.Config, chainAConnID string) error
 	// CompleteIBCChannelHandshake completes the channel handshake in cases when ChanOpenInit was initiated
 	// by some transaction that was previously executed on the chain. For example,
 	// ICA MsgRegisterInterchainAccount will perform ChanOpenInit during its execution.