+++ testutil/keeper/btcstaking.go
@@ -20,6 +20,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	appparams "github.com/babylonlabs-io/babylon/v3/app/params"
+	"github.com/babylonlabs-io/babylon/v3/testutil/mocks"
 	"github.com/babylonlabs-io/babylon/v3/x/btcstaking/keeper"
 	"github.com/babylonlabs-io/babylon/v3/x/btcstaking/types"
 	bsckeeper "github.com/babylonlabs-io/babylon/v3/x/btcstkconsumer/keeper"
@@ -34,8 +35,9 @@ func BTCStakingKeeperWithStore(
 	btclcKeeper types.BTCLightClientKeeper,
 	btccKeeper types.BtcCheckpointKeeper,
 	iKeeper types.IncentiveKeeper,
+	chKeeper *mocks.MockZoneConciergeChannelKeeper,
 ) (*keeper.Keeper, sdk.Context) {
-	return BTCStakingKeeperWithStoreAndBank(t, db, stateStore, storeKey, btclcKeeper, btccKeeper, iKeeper, nil)
+	return BTCStakingKeeperWithStoreAndBank(t, db, stateStore, storeKey, btclcKeeper, btccKeeper, iKeeper, nil, chKeeper)
 }
 
 func BTCStakingKeeperWithStoreAndBank(
@@ -47,6 +49,7 @@ func BTCStakingKeeperWithStoreAndBank(
 	btccKeeper types.BtcCheckpointKeeper,
 	iKeeper types.IncentiveKeeper,
 	bankKeeper types.BankKeeper,
+	chKeeper *mocks.MockZoneConciergeChannelKeeper,
 ) (*keeper.Keeper, sdk.Context) {
 	if storeKey == nil {
 		storeKey = storetypes.NewKVStoreKey(types.StoreKey)
@@ -67,7 +70,7 @@ func BTCStakingKeeperWithStoreAndBank(
 		cdc,
 		runtime.NewKVStoreService(bscStoreKey),
 		nil,
-		nil,
+		chKeeper,
 		nil,
 		authtypes.NewModuleAddress(govtypes.ModuleName).String(),
 	)
@@ -97,8 +100,9 @@ func BTCStakingKeeper(
 	btclcKeeper types.BTCLightClientKeeper,
 	btccKeeper types.BtcCheckpointKeeper,
 	iKeeper types.IncentiveKeeper,
+	chKeeper *mocks.MockZoneConciergeChannelKeeper,
 ) (*keeper.Keeper, sdk.Context) {
-	return BTCStakingKeeperWithStoreKey(t, nil, btclcKeeper, btccKeeper, iKeeper)
+	return BTCStakingKeeperWithStoreKey(t, nil, btclcKeeper, btccKeeper, iKeeper, chKeeper)
 }
 
 func BTCStakingKeeperWithStoreKey(
@@ -107,11 +111,12 @@ func BTCStakingKeeperWithStoreKey(
 	btclcKeeper types.BTCLightClientKeeper,
 	btccKeeper types.BtcCheckpointKeeper,
 	iKeeper types.IncentiveKeeper,
+	chKeeper *mocks.MockZoneConciergeChannelKeeper,
 ) (*keeper.Keeper, sdk.Context) {
 	db := dbm.NewMemDB()
 	stateStore := store.NewCommitMultiStore(db, log.NewTestLogger(t), storemetrics.NewNoOpMetrics())
 
-	k, ctx := BTCStakingKeeperWithStore(t, db, stateStore, storeKey, btclcKeeper, btccKeeper, iKeeper)
+	k, ctx := BTCStakingKeeperWithStore(t, db, stateStore, storeKey, btclcKeeper, btccKeeper, iKeeper, chKeeper)
 
 	// Initialize params
 	if err := k.SetParams(ctx, types.DefaultParams()); err != nil {