+++ x/zoneconcierge/keeper/genesis.go
@@ -8,7 +8,7 @@ import (
 
 // InitGenesis initializes the keeper state from a provided initial genesis state.
 func (k Keeper) InitGenesis(ctx context.Context, gs types.GenesisState) error {
-	if err := k.SetPort(ctx, gs.PortId); err != nil {
+	if err := k.channelKeeper.InitGenesis(ctx, gs); err != nil {
 		return err
 	}
 
@@ -54,7 +54,7 @@ func (k Keeper) ExportGenesis(ctx context.Context) (*types.GenesisState, error) 
 
 	return &types.GenesisState{
 		Params:             k.GetParams(ctx),
-		PortId:             k.GetPort(ctx),
+		PortId:             k.channelKeeper.GetPort(ctx),
 		FinalizedHeaders:   fh,
 		SealedEpochsProofs: se,
 		BsnBtcStates:       cs,